baseline_comparison/requirements.txt
openai>=1.30.0
anthropic>=0.30.0
python-docx>=1.1.0
python-dotenv>=1.0.0
